refactoring: move api-specific tests to api files
